# src/main/resources/openapi/notification-service.yaml
openapi: 3.1.0
info:
  title: Notification Service
  description: Shared platform service for email and webhook notifications
  version: 1.0.0
  contact:
    name: Platform Team
servers:
  - url: /api/v1
paths:
  /notifications/send:
    post:
      operationId: sendNotification
      summary: Send a single notification
      description: Submit a notification for async delivery via email or webhook
      tags:
        - notifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendNotificationRequest'
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: Notification queued for delivery
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendNotificationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /notifications/send-bulk:
    post:
      operationId: sendBulkNotifications
      summary: Send multiple notifications in a single request
      description: Submit a batch of notifications for async delivery
      tags:
        - notifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkSendNotificationRequest'
      responses:
        '207':
          description: Multi-status response with per-item results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkSendNotificationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /notification-templates:
    get:
      operationId: listTemplates
      summary: List notification templates
      description: Returns a paginated list of notification templates
      tags:
        - notification-templates
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Filter'
      responses:
        '200':
          description: Paginated list of templates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createTemplate
      summary: Create a notification template
      tags:
        - notification-templates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTemplateRequest'
      responses:
        '201':
          description: Template created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /notification-templates/{templateId}:
    get:
      operationId: getTemplate
      summary: Get a template by ID
      tags:
        - notification-templates
      parameters:
        - $ref: '#/components/parameters/TemplateId'
      responses:
        '200':
          description: Template details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      operationId: updateTemplate
      summary: Update a template
      tags:
        - notification-templates
      parameters:
        - $ref: '#/components/parameters/TemplateId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTemplateRequest'
      responses:
        '200':
          description: Template updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    delete:
      operationId: deleteTemplate
      summary: Soft-delete a template
      tags:
        - notification-templates
      parameters:
        - $ref: '#/components/parameters/TemplateId'
      responses:
        '204':
          description: Template deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: Idempotency key to prevent duplicate sends
    TemplateId:
      name: templateId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    Sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
      description: Sort expression (e.g., createdAt,desc)
    Filter:
      name: filter
      in: query
      required: false
      schema:
        type: string
      description: Filter expression (e.g., status:eq:DELIVERED)
  schemas:
    Channel:
      type: string
      enum:
        - EMAIL
        - WEBHOOK
    DeliveryStatus:
      type: string
      enum:
        - QUEUED
        - PROCESSING
        - DELIVERED
        - FAILED
        - DRY_RUN
    Recipient:
      type: object
      required:
        - address
      properties:
        address:
          type: string
          format: email
        name:
          type: string
    TemplateVariable:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
        type:
          type: string
          enum:
            - STRING
            - NUMBER
            - DATE
            - BOOLEAN
            - LIST
            - OBJECT
        required:
          type: boolean
          default: true
    SendNotificationRequest:
      type: object
      required:
        - channel
      properties:
        channel:
          $ref: '#/components/schemas/Channel'
        recipient:
          $ref: '#/components/schemas/Recipient'
        templateId:
          type: string
        templateVariables:
          type: object
          additionalProperties: true
        subject:
          type: string
        body:
          type: string
        contentType:
          type: string
          default: text/html
        webhookSubscriptionId:
          type: string
        payload:
          type: object
          additionalProperties: true
        callbackUrl:
          type: string
          format: uri
        metadata:
          type: object
          additionalProperties:
            type: string
    SendNotificationResponse:
      type: object
      properties:
        deliveryId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/DeliveryStatus'
        statusUrl:
          type: string
        queuedAt:
          type: string
          format: date-time
    BulkSendNotificationRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BulkSendItem'
          minItems: 1
          maxItems: 100
    BulkSendItem:
      type: object
      required:
        - clientItemId
        - data
      properties:
        clientItemId:
          type: string
        data:
          $ref: '#/components/schemas/SendNotificationRequest'
    BulkSendNotificationResponse:
      type: object
      properties:
        summary:
          $ref: '#/components/schemas/BulkSummary'
        results:
          type: array
          items:
            $ref: '#/components/schemas/BulkItemResult'
    BulkSummary:
      type: object
      properties:
        total:
          type: integer
        succeeded:
          type: integer
        failed:
          type: integer
    BulkItemResult:
      type: object
      properties:
        clientItemId:
          type: string
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/SendNotificationResponse'
        error:
          $ref: '#/components/schemas/ProblemDetail'
    CreateTemplateRequest:
      type: object
      required:
        - name
        - channel
        - bodyTemplate
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        channel:
          $ref: '#/components/schemas/Channel'
        subject:
          type: string
          maxLength: 998
        bodyTemplate:
          type: string
          minLength: 1
        variables:
          type: array
          items:
            $ref: '#/components/schemas/TemplateVariable'
    UpdateTemplateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        subject:
          type: string
          maxLength: 998
        bodyTemplate:
          type: string
          minLength: 1
        variables:
          type: array
          items:
            $ref: '#/components/schemas/TemplateVariable'
    TemplateResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        channel:
          $ref: '#/components/schemas/Channel'
        subject:
          type: string
        bodyTemplate:
          type: string
        variables:
          type: array
          items:
            $ref: '#/components/schemas/TemplateVariable'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TemplateListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/TemplateResponse'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
...
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Forbidden:
      description: Forbidden
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Conflict:
      description: Conflict
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    UnprocessableEntity:
      description: Validation error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'